new problems

+++ java_sol/AppleDis.java
@@ -8,16 +8,16 @@ public class AppleDis {
         int n = sc.nextInt();
 
         int[] aps = new int[n];
-        int sm = 0;
+        long sm = 0;
 
         for (int i = 0; i < n; i++) {
             aps[i] = sc.nextInt();
             sm += aps[i];
         }
 
-        int ans = sm;
+        long ans = sm;
         for (int i = 0; i < (1 << n); i++) {
-            int cs = 0;
+            long cs = 0;
             for (int j = 0; j < n; j++) {
                 if (((i >> j) & 1) == 1) {
                     cs += aps[j];
